+++ src/connector/worker.js
@@ -74,10 +74,12 @@ export default class Worker {
           })
           .then((jobRes) => {
             callback(null, jobRes);
-          }, (err) => {
+          })
+          .catch((err) => {
             req.hull.metric.increment(`ship.job.${jobName}.error`);
             this.instrumentation.catchError(err, {
-              job_id: job.id
+              job_id: job.id,
+              job_payload: jobData
             }, {
               job_name: job.data.name,
               organization: _.get(job.data.context, "query.organization"),