+++ src/utils/exit-handler.js
@@ -3,16 +3,18 @@
  */
 export default function exitHandler(promise) {
   function exitNow() {
-    console.warn("exitHandler.exitNow");
+    console.warn("connector.exitHandler.exitNow");
     process.exit(0);
   }
 
   function handleExit() {
-    console.log("exitHandler.handleExit");
-    setTimeout(exitNow, 30000);
+    const waiting = 30000;
+    console.log("connector.exitHandler.handleExit", { waiting });
+    setTimeout(exitNow, waiting);
     promise().then(exitNow, exitNow);
   }
 
   process.on("SIGINT", handleExit);
   process.on("SIGTERM", handleExit);
+  process.on("gracefulExit", handleExit);
 }