+++ tests/connector/hull-connector.js
@@ -6,6 +6,11 @@ import HullConnector from "../../src/connector/hull-connector";
 import HullStub from "../support/hull-stub";
 
 describe("HullConnector", () => {
+
+  after(() => {
+    process.removeAllListeners("exit");
+  });
+
   it("should return an object of functions", () => {
     const connector = new HullConnector(HullStub);
     expect(connector).to.be.object;
@@ -93,5 +98,34 @@ describe("HullConnector", () => {
 
     expect(processSpy.calledOnce).to.be.true;
     expect(processSpy.getCall(0).args[0]).to.be.equal("queueApp");
+  });
+
+  it("should allow to setup custom middleware at the end of pre-handler middleware stack", () => {
+    const appStub = {
+      use: () => {},
+      engine: () => {},
+      set: () => {}
+    };
+
+    const workerStub = {
+      use: () => {},
+      setJobs: () => {}
+    };
+
+    const appUseSpy = sinon.spy(appStub, "use");
+    const workerUseSpy = sinon.spy(workerStub, "use");
+
+    const customMiddleware = (req, res, next) => {};
+    const connector = new HullConnector(HullStub);
+    connector.use(customMiddleware);
+    connector.setupApp(appStub);
+    connector._worker = workerStub;
+    connector.worker({});
+
+    expect(appUseSpy.called).to.be.true;
+    expect(appUseSpy.lastCall.args[0]).to.be.eql(customMiddleware);
+
+    expect(workerUseSpy.called).to.be.true;
+    expect(workerUseSpy.lastCall.args[0]).to.be.eql(customMiddleware);
   });
 });