+++ tests/infra/instrumentation-tests.js
@@ -10,5 +10,22 @@ describe("Instrumentation", () => {
     const instrumentation = new Instrumentation();
     expect(instrumentation).to.be.an("object");
     delete process.env.SENTRY_URL;
+    instrumentation.raven.uninstall();
+    process.removeAllListeners("gracefulExit");
+  });
+
+  it("should handle unhandled rejection with undefined rejection reason", (done) => {
+    process.on("gracefulExit", () => {
+      done();
+      instrumentation.raven.uninstall();
+      process.removeAllListeners("gracefulExit");
+    });
+    process.env.SENTRY_URL = "https://user:pass@sentry.io/138436";
+    const instrumentation = new Instrumentation({ exitOnError: true });
+    expect(instrumentation).to.be.an("object");
+    delete process.env.SENTRY_URL;
+    new Promise((resolve, reject) => {
+      reject();
+    });
   });
 });