+++ tests/support/hull-stub.js
@@ -3,7 +3,7 @@ import Promise from "bluebird";
 
 export default class HullStub {
   constructor() {
-    this.id = _.uniqueId('ship-');
+    this.id = _.uniqueId("ship-");
     this.logger = {
       info: console.log, //() {},
       debug: console.log, //() {}
@@ -16,7 +16,7 @@ export default class HullStub {
   post(id) { return Promise.resolve({ id }); }
 
   configuration() {
-    return { id: this.id , secret: "shutt", organization: "xxx.hulltest.rocks" };
+    return { id: this.id, secret: "shutt", organization: "xxx.hulltest.rocks" };
   }
 
   static Middleware() {
@@ -27,6 +27,6 @@ export default class HullStub {
 }
 
 HullStub.logger = {
-  info: console.log, //() {},
-  debug: console.log, //() {}
+  info: console.log, // () {},
+  debug: console.log, // () {}
 };