+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+# 0.12.0
+* replaces low level api client with a separate library
+* **breaking**: `client.utils.extract.handle` replaced by `ctx.helpers.handleExtract`
+* **breaking**: `client.utils.extract.request` replaced by `ctx.helpers.requestExtract`
+* **breaking**: removed `lib/utils/batchHandler` in favour of `lib/utils/notifHandler`
+* **breaking**: removed `service` param on `Hull.Connector` in favor of custom middleware using `.use` method
+* **breaking**: `QueueAgent` instance now accepts `QueueAdapter` instance instead of it's name. Load you adapter now like this:
+  ```js
+  import BullAdapter from "hull/lib/infra/queue/adapter/bull";
+
+  const queueAdapter = new BullAdapter(options);
+
+  const queue = new Queue(queueAdapter);
+  ```
+* don't exit on unhandled errors in by default, turn it on for workers
+* handle unhandled rejection which was rejected to an undefined value
+* add more context to the express app requests middleware stack
+* adds logging to all oauth handler steps
+* upgrade hull-client to v1.1.1 to add support for `scopes` claim in auth tokens
+
 # 0.11.12
 * when handling batch extract, `notifHandler` should respond as soon as we have started to download and process JSON file. Otherwise in case of big extract files in may lead to response timeout